changed the db schema data types
there were some issues with the data types when creating the tables.
zip codes were having their leading 0s dropped.  i added a zerofill
to sovle that issue.  Also there was a transaction amount in the
db that was 5.5600000005 but it wasn't present in the python df so
I added double(7,2) the 2 is for 2 decimal places.